add initial test file and move notificationButtonProvider
