+++ examples/test-site/src/data/pages/notification/index.tsx
@@ -42,7 +42,6 @@ const NotificationViewer = () => {
 
 const ChildWithNotifications = asBodiless(() => {
   const { notifications, setNotifications } = useNotifyFromNode();
-
   const addRandomNotification = useCallback(
     () => {
       const id = v1();
