+++ packages/bodiless-core/__tests__/NotificationProvider.test.tsx
@@ -1,0 +1,36 @@
+/**
+ * Copyright © 2020 Johnson & Johnson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { mount } from 'enzyme';
+import React, { useContext } from 'react';
+import { NotificationContext, useNotify } from '../src/NotificationProvider';
+
+const mockNotifications = [
+  {
+    id: '0',
+    message: 'foo',
+  },
+];
+const ChildWithNotifications = () => {
+  useNotify(mockNotifications);
+  return null;
+};
+describe(' useNotify', () => {
+  it('renders the notifications', () => {
+    const Wrapper = mount(
+      <ChildWithNotifications />,
+    );
+    console.log(Wrapper.debug());
+  });
+});
